requirements.txt
aiohttp==3.7.3
altgraph==0.17
asgiref==3.2.3
astroid==2.4.1
async-timeout==3.0.1
attrs==20.3.0
auto-py-to-exe==2.7.8
blessings==1.7
bottle==0.12.18
bottle-websocket==0.2.9
bpython==0.19
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9
cssselect2==0.3.0
curtsies==0.3.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.5
django-admin-interface==0.16.1
django-braces==1.14.0
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-colorfield==0.4.1
django-crispy-forms==1.9.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-private-chat==0.3.0
django-summernote==0.8.11.6
django-taggit==1.3.0
django-tinymce==3.0.1
djangorestframework==3.11.0
Eel==0.12.4
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
gunicorn==20.0.4
html5lib==1.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.6.2
mccabe==0.6.1
multidict==5.1.0
oauthlib==3.1.0
pefile==2019.4.18
Pillow==7.1.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.6.1
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
PyJWT==1.7.1
pylint==2.5.2
Pyphen==0.9.5
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
pywin32-ctypes==0.2.0
redis==2.10.6
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.0
tinycss2==1.0.2
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.25.8
validate-email==1.3
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
websockets==8.1
whichcraft==0.6.1
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.1.2
